bot/requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
sqlalchemy>=2.0.36
aiosqlite==0.19.0

